outlook_mcp/requirements.txt
fastmcp 
requests
pydantic
mcp[cli]
google-generativeai
openai
python-dotenv
websockets